@@ -6,8 +6,8 @@ import (
 )
 
 type Message struct {
-	Text       string
 	Expiry     time.Time
 	Foreground color.Color
 	Background color.Color
+	Text       string
 }
