@@ -21,31 +21,31 @@ const (
 )
 
 type Buffer struct {
-	lines                 []Line
-	savedCursorPos        Position
-	savedCursorAttr       *CellAttributes
-	cursorShape           CursorShape
-	savedCharsets         []*map[rune]rune
-	savedCurrentCharset   int
-	topMargin             uint // see DECSTBM docs - this is for scrollable regions
-	bottomMargin          uint // see DECSTBM docs - this is for scrollable regions
-	viewWidth             uint16
-	viewHeight            uint16
-	cursorPosition        Position // raw
-	cursorAttr            CellAttributes
-	scrollLinesFromBottom uint
-	maxLines              uint64
-	tabStops              []uint16
-	charsets              []*map[rune]rune // array of 2 charsets, nil means ASCII (no conversion)
-	currentCharset        int              // active charset index in charsets array, valid values are 0 or 1
-	modes                 Modes
-	selectionStart        *Position
-	selectionEnd          *Position
-	highlightStart        *Position
 	highlightEnd          *Position
+	selectionStart        *Position
+	savedCursorAttr       *CellAttributes
+	selectionEnd          *Position
 	highlightAnnotation   *Annotation
+	highlightStart        *Position
+	cursorAttr            CellAttributes
 	sixels                []Sixel
+	lines                 []Line
+	savedCharsets         []*map[rune]rune
+	tabStops              []uint16
+	charsets              []*map[rune]rune
+	cursorPosition        Position
+	savedCursorPos        Position
+	topMargin             uint
+	maxLines              uint64
+	currentCharset        int
+	scrollLinesFromBottom uint
+	bottomMargin          uint
+	savedCurrentCharset   int
 	selectionMu           sync.Mutex
+	viewHeight            uint16
+	viewWidth             uint16
+	modes                 Modes
+	cursorShape           CursorShape
 }
 
 type Annotation struct {
@@ -126,7 +126,6 @@ func (buffer *Buffer) getAreaScrollRange() (top uint64, bottom uint64) {
 }
 
 func (buffer *Buffer) areaScrollDown(lines uint16) {
-
 	// NOTE: bottom is exclusive
 	top, bottom := buffer.getAreaScrollRange()
 
@@ -141,7 +140,6 @@ func (buffer *Buffer) areaScrollDown(lines uint16) {
 }
 
 func (buffer *Buffer) areaScrollUp(lines uint16) {
-
 	// NOTE: bottom is exclusive
 	top, bottom := buffer.getAreaScrollRange()
 
@@ -281,7 +279,6 @@ func (buffer *Buffer) deleteLine() {
 }
 
 func (buffer *Buffer) insertLine() {
-
 	if !buffer.InScrollableRegion() {
 		pos := buffer.RawLine()
 		maxLines := buffer.GetMaxLines()
@@ -322,7 +319,6 @@ func (buffer *Buffer) insertLine() {
 }
 
 func (buffer *Buffer) insertBlankCharacters(count int) {
-
 	index := int(buffer.RawLine())
 	for i := 0; i < count; i++ {
 		cells := buffer.lines[index].cells
@@ -331,7 +327,6 @@ func (buffer *Buffer) insertBlankCharacters(count int) {
 }
 
 func (buffer *Buffer) insertLines(count int) {
-
 	if buffer.HasScrollableRegion() && !buffer.InScrollableRegion() {
 		// should have no effect outside of scrollable region
 		return
@@ -342,11 +337,9 @@ func (buffer *Buffer) insertLines(count int) {
 	for i := 0; i < count; i++ {
 		buffer.insertLine()
 	}
-
 }
 
 func (buffer *Buffer) deleteLines(count int) {
-
 	if buffer.HasScrollableRegion() && !buffer.InScrollableRegion() {
 		// should have no effect outside of scrollable region
 		return
@@ -357,11 +350,9 @@ func (buffer *Buffer) deleteLines(count int) {
 	for i := 0; i < count; i++ {
 		buffer.deleteLine()
 	}
-
 }
 
 func (buffer *Buffer) index() {
-
 	// This sequence causes the active position to move downward one line without changing the column position.
 	// If the active position is at the bottom margin, a scroll up is performed."
 
@@ -390,7 +381,6 @@ func (buffer *Buffer) index() {
 }
 
 func (buffer *Buffer) reverseIndex() {
-
 	cursorVY := buffer.convertRawLineToViewLine(buffer.cursorPosition.Line)
 
 	if uint(cursorVY) == buffer.topMargin {
@@ -402,7 +392,6 @@ func (buffer *Buffer) reverseIndex() {
 
 // write will write a rune to the terminal at the position of the cursor, and increment the cursor position
 func (buffer *Buffer) write(runes ...MeasuredRune) {
-
 	// scroll to bottom on input
 	buffer.scrollLinesFromBottom = 0
 
@@ -451,7 +440,6 @@ func (buffer *Buffer) write(runes ...MeasuredRune) {
 				// no more room on line and wrapping is disabled
 				return
 			}
-
 		} else {
 
 			for int(buffer.CursorColumn()) >= len(line.cells) {
@@ -483,7 +471,6 @@ func (buffer *Buffer) inDoWrap() bool {
 }
 
 func (buffer *Buffer) backspace() {
-
 	if buffer.cursorPosition.Col == 0 {
 		line := buffer.getCurrentLine()
 		if line.wrapped {
@@ -498,7 +485,6 @@ func (buffer *Buffer) backspace() {
 }
 
 func (buffer *Buffer) carriageReturn() {
-
 	cursorVY := buffer.convertRawLineToViewLine(buffer.cursorPosition.Line)
 
 	for {
@@ -517,7 +503,6 @@ func (buffer *Buffer) carriageReturn() {
 }
 
 func (buffer *Buffer) tab() {
-
 	tabStop := buffer.getNextTabStopAfter(buffer.cursorPosition.Col)
 	for buffer.cursorPosition.Col < tabStop && buffer.cursorPosition.Col < buffer.viewWidth-1 { // @todo rightMargin
 		buffer.write(MeasuredRune{Rune: ' ', Width: 1})
@@ -526,7 +511,6 @@ func (buffer *Buffer) tab() {
 
 // return next tab stop x pos
 func (buffer *Buffer) getNextTabStopAfter(col uint16) uint16 {
-
 	defaultStop := col + (TabSize - (col % TabSize))
 	if defaultStop == col {
 		defaultStop += TabSize
@@ -565,7 +549,6 @@ func (buffer *Buffer) verticalTab() {
 }
 
 func (buffer *Buffer) newLineEx(forceCursorToMargin bool) {
-
 	if buffer.IsNewLineMode() || forceCursorToMargin {
 		buffer.cursorPosition.Col = 0
 	}
@@ -581,7 +564,6 @@ func (buffer *Buffer) newLineEx(forceCursorToMargin bool) {
 }
 
 func (buffer *Buffer) movePosition(x int16, y int16) {
-
 	var toX uint16
 	var toY uint16
 
@@ -602,7 +584,6 @@ func (buffer *Buffer) movePosition(x int16, y int16) {
 }
 
 func (buffer *Buffer) setPosition(col uint16, line uint16) {
-
 	useCol := col
 	useLine := line
 	maxLine := buffer.ViewHeight() - 1
@@ -652,7 +633,6 @@ func (buffer *Buffer) getCurrentLine() *Line {
 }
 
 func (buffer *Buffer) getViewLine(index uint16) *Line {
-
 	if index >= buffer.ViewHeight() {
 		return &buffer.lines[len(buffer.lines)-1]
 	}
@@ -672,7 +652,6 @@ func (buffer *Buffer) getViewLine(index uint16) *Line {
 }
 
 func (buffer *Buffer) eraseLine() {
-
 	buffer.clearSixelsAtRawLine(buffer.cursorPosition.Line)
 
 	line := buffer.getCurrentLine()
@@ -720,7 +699,6 @@ func (buffer *Buffer) eraseDisplay() {
 }
 
 func (buffer *Buffer) deleteChars(n int) {
-
 	line := buffer.getCurrentLine()
 	if int(buffer.cursorPosition.Col) >= len(line.cells) {
 		return
@@ -734,7 +712,6 @@ func (buffer *Buffer) deleteChars(n int) {
 }
 
 func (buffer *Buffer) eraseCharacters(n int) {
-
 	line := buffer.getCurrentLine()
 
 	max := int(buffer.cursorPosition.Col) + n
