@@ -124,7 +124,6 @@ func (t *Terminal) handleCSI(readChan chan MeasuredRune) (renderRequired bool) {
 	_ = raw
 	t.log("UNKNOWN CSI P(%s) I(%s) %c", strings.Join(params, ";"), string(intermediate), final)
 	return false
-
 }
 
 type WindowState uint8
@@ -160,7 +159,6 @@ type WindowManipulator interface {
 }
 
 func (t *Terminal) csiWindowManipulation(params []string) (renderRequired bool) {
-
 	if t.windowManipulator == nil {
 		return false
 	}
@@ -171,7 +169,7 @@ func (t *Terminal) csiWindowManipulation(params []string) (renderRequired bool) 
 			t.windowManipulator.Restore()
 		case "2":
 			t.windowManipulator.Minimise()
-		case "3": //move window
+		case "3": // move window
 			if i+2 >= len(params) {
 				return false
 			}
@@ -179,7 +177,7 @@ func (t *Terminal) csiWindowManipulation(params []string) (renderRequired bool) 
 			y, _ := strconv.Atoi(params[i+2])
 			i += 2
 			t.windowManipulator.Move(x, y)
-		case "4": //resize h,w
+		case "4": // resize h,w
 			w, h := t.windowManipulator.SizeInPixels()
 			if i+1 < len(params) {
 				h, _ = strconv.Atoi(params[i+1])
@@ -303,7 +301,6 @@ func (t *Terminal) csiWindowManipulation(params []string) (renderRequired bool) 
 // CSI c
 // Send Device Attributes (Primary/Secondary/Tertiary DA)
 func (t *Terminal) csiSendDeviceAttributesHandler(params []string) (renderRequired bool) {
-
 	// we are VT100
 	// for DA1 we'll respond ?1;2
 	// for DA2 we'll respond >0;0;0
@@ -320,7 +317,6 @@ func (t *Terminal) csiSendDeviceAttributesHandler(params []string) (renderRequir
 // CSI n
 // Device Status Report (DSR)
 func (t *Terminal) csiDeviceStatusReportHandler(params []string) (renderRequired bool) {
-
 	if len(params) == 0 {
 		return false
 	}
@@ -405,7 +401,6 @@ func (t *Terminal) csiCursorBackwardHandler(params []string) (renderRequired boo
 // CSI E
 // Cursor Next Line Ps Times (default = 1) (CNL)
 func (t *Terminal) csiCursorNextLineHandler(params []string) (renderRequired bool) {
-
 	distance := 1
 	if len(params) > 0 {
 		var err error
@@ -423,7 +418,6 @@ func (t *Terminal) csiCursorNextLineHandler(params []string) (renderRequired boo
 // CSI F
 // Cursor Preceding Line Ps Times (default = 1) (CPL)
 func (t *Terminal) csiCursorPrecedingLineHandler(params []string) (renderRequired bool) {
-
 	distance := 1
 	if len(params) > 0 {
 		var err error
@@ -667,7 +661,6 @@ func (t *Terminal) csiSetModes(modes []string, enabled bool) bool {
 }
 
 func parseModes(mode string) []string {
-
 	var output []string
 
 	if mode == "" {
@@ -689,9 +682,7 @@ func parseModes(mode string) []string {
 }
 
 func (t *Terminal) csiSetMode(modes string, enabled bool) bool {
-
 	for _, modeStr := range parseModes(modes) {
-
 		switch modeStr {
 		case "4":
 			t.activeBuffer.modes.ReplaceMode = !enabled
@@ -717,7 +708,7 @@ func (t *Terminal) csiSetMode(modes string, enabled bool) bool {
 			t.activeBuffer.modes.OriginMode = enabled
 		case "?7":
 			// auto-wrap mode
-			//DECAWM
+			// DECAWM
 			t.activeBuffer.modes.AutoWrap = enabled
 		case "?9":
 			if enabled {
@@ -874,14 +865,13 @@ func (t *Terminal) csiEraseInDisplayHandler(params []string) (renderRequired boo
 // CSI K
 // Erase in Line (EL), VT100
 func (t *Terminal) csiEraseInLineHandler(params []string) (renderRequired bool) {
-
 	n := "0"
 	if len(params) > 0 {
 		n = params[0]
 	}
 
 	switch n {
-	case "0", "": //erase adter cursor
+	case "0", "": // erase adter cursor
 		t.GetActiveBuffer().eraseLineFromCursor()
 	case "1": // erase to cursor inclusive
 		t.GetActiveBuffer().eraseLineToCursor()
@@ -896,14 +886,13 @@ func (t *Terminal) csiEraseInLineHandler(params []string) (renderRequired bool) 
 // CSI m
 // Character Attributes (SGR)
 func (t *Terminal) sgrSequenceHandler(params []string) bool {
-
 	if len(params) == 0 {
 		params = []string{"0"}
 	}
 
 	for i := range params {
 
-		p := strings.Replace(strings.Replace(params[i], "[", "", -1), "]", "", -1)
+		p := strings.ReplaceAll(strings.ReplaceAll(params[i], "[", ""), "]", "")
 
 		switch p {
 		case "00", "0", "":
